+++ cleverpush/src/main/java/com/cleverpush/CleverPushPreferences.java
@@ -9,5 +9,7 @@ public class CleverPushPreferences {
     public static final String SUBSCRIPTION_LAST_SYNC = "subscriptionLastSync";
     public static final String SUBSCRIPTION_TAGS = "subscriptionTags";
     public static final String SUBSCRIPTION_ATTRIBUTES = "subscriptionAttributes";
+    public static final String SUBSCRIPTION_LANGUAGE = "subscriptionLanguage";
+    public static final String SUBSCRIPTION_COUNTRY = "subscriptionCountry";
 
 }