+++ cleverpush/src/main/java/com/cleverpush/NotificationOpenedActivity.java
@@ -6,17 +6,19 @@ import android.os.Bundle;
 
 public class NotificationOpenedActivity extends Activity {
 
-   @Override
-   protected void onCreate(Bundle savedInstanceState) {
-      super.onCreate(savedInstanceState);
-      NotificationOpenedProcessor.processIntent(this, getIntent());
-      finish();
-   }
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        System.out.println("CleverPush: NotificationOpenedActivity onCreate");
+        super.onCreate(savedInstanceState);
+        NotificationOpenedProcessor.processIntent(this, getIntent());
+        finish();
+    }
 
-   @Override
-   protected void onNewIntent(Intent intent) {
-      super.onNewIntent(intent);
-      NotificationOpenedProcessor.processIntent(this, getIntent());
-      finish();
-   }
+    @Override
+    protected void onNewIntent(Intent intent) {
+        System.out.println("CleverPush: NotificationOpenedActivity onNewIntent");
+        super.onNewIntent(intent);
+        NotificationOpenedProcessor.processIntent(this, getIntent());
+        finish();
+    }
 }